rtsp: Use composition instead of inheritance
Relatest to #1078
